update agent tools
