+++ code/chat_agent/Dockerfile
@@ -6,6 +6,7 @@ RUN  pip3 install -r requirements.txt --target "${LAMBDA_TASK_ROOT}"
 
 # Copy function code
 COPY aws_agent.py ${LAMBDA_TASK_ROOT}
+COPY tools ${LAMBDA_TASK_ROOT}/tools
 
 # Set the CMD to your handler (could also be done as a parameter override outside of the Dockerfile)
 CMD [ "aws_agent.lambda_handler" ]
